+++ ignite/cmd/network_request.go
@@ -17,6 +17,8 @@ func NewNetworkRequest() *cobra.Command {
 		NewNetworkRequestReject(),
 		NewNetworkRequestVerify(),
 		NewNetworkRequestAddAccount(),
+		NewNetworkRequestRemoveAccount(),
+		NewNetworkRequestRemoveValidator(),
 	)
 
 	return c